// packages/backend/src/_common/model/structure-doc/StructureCustomDoc.type.ts
export type StructureCustomDoc = {
  // DATES UTILES
  DATE_JOUR: string;
  DATE_JOUR_HEURE: string;
  DATE_JOUR_LONG: string;
  // RESPONSABLE STRUCTURE
  RESPONSABLE_NOM: string;
  RESPONSABLE_PRENOM: string;
  RESPONSABLE_FONCTION: string;

  // INFOS STRUCTURE
  STRUCTURE_NOM: string;
  STRUCTURE_TYPE: string;
  STRUCTURE_ADRESSE: string;
  STRUCTURE_COMPLEMENT_ADRESSE: string;
  STRUCTURE_VILLE: string;
  STRUCTURE_CODE_POSTAL: string;

  // SI ADRESSE COURRIER DIFFERENTE
  STRUCTURE_COURRIER_ADRESSE: string;
  STRUCTURE_COURRIER_COMPLEMENT_ADRESSE: string;
  STRUCTURE_COURRIER_VILLE: string;
  STRUCTURE_COURRIER_CODE_POSTAL: string;
  //
  USAGER_REF: string;
  USAGER_CUSTOM_REF: string;
  USAGER_CIVILITE: string;
  USAGER_NOM: string;
  USAGER_PRENOM: string;
  USAGER_SURNOM: string;
  USAGER_DATE_NAISSANCE: string;
  USAGER_LIEU_NAISSANCE: string;
  USAGER_PHONE: string;
  USAGER_EMAIL: string;
  //
  STATUT_DOM: string;
  TYPE_DOM: string;
  DATE_DEBUT_DOM: string;
  DATE_FIN_DOM: string;
  DATE_PREMIERE_DOM: string;
  DATE_DERNIER_PASSAGE: string;

  //  RADIATIONS
  DATE_RADIATION: string;
  MOTIF_RADIATION: string;

  // ENTRETIEN
  ENTRETIEN_ORIENTE_PAR: string;
  ENTRETIEN_DOMICILIATION_EXISTANTE: string;
  ENTRETIEN_REVENUS: string;
  ENTRETIEN_LIEN_COMMUNE: string;
  ENTRETIEN_COMPOSITION_MENAGE: string;
  ENTRETIEN_SITUATION_RESIDENTIELLE: string;
  ENTRETIEN_CAUSE_INSTABILITE: string;
  ENTRETIEN_RAISON_DEMANDE: string;
  ENTRETIEN_ACCOMPAGNEMENT: string;
};